+++ src/utils/listeners/scroll.js
@@ -14,13 +14,17 @@ const useScrollListener = () => {
     const scrollFunction = () => {
         const documentScrollTop = document.body.scrollTop || document.documentElement.scrollTop;
 
-        sectionsList.map((section, index) => {
-            if(documentScrollTop <= (sectionRefs.current[0].offsetTop - (windowHeight / 2.5)))
-                return setActiveSection('Overview')
+        sectionsList
+            .filter(section => !section.isStandalone)
+            .map((section, index) => {
+                if(window.location.pathname !== "/") return;
 
-            if(documentScrollTop >= (sectionRefs.current[index].offsetTop - (windowHeight / 2.5)))
-                return setActiveSection(section.name)
-        });
+                if(documentScrollTop <= (sectionRefs.current[0].offsetTop - (windowHeight / 2.5)))
+                    return setActiveSection('Overview')
+
+                if(documentScrollTop >= (sectionRefs.current[index].offsetTop - (windowHeight / 2.5)))
+                    return setActiveSection(section.name)
+            });
 
         if (documentScrollTop > 500)
             setRevealScrollTop(revealScrollTop ? false : true)